cyber_job_hunter/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
fastapi-users[sqlalchemy]>=15.0.0
fastapi-users-db-sqlalchemy>=7.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
httpx>=0.27.0
resend>=2.0.0
pdfplumber>=0.11.0
python-docx>=1.0.0
